Add support for injector building
Having a way to split the configuration of the injector allows for
sharing of configurations and promotes code reuse.

Providing a configuration for creating an injector container reduces the
amount of boilerplate code required to use this package.

Providing a configuration for loading service definitions allows use
with Zend components.